Add info tags to define PVI structure from controllers
Refactor helper functions to preserve PV prefix until later so that PVI
PV can use it.
Remove call to SetDeviceName and construct full PVs when creating
records.
Add call to create info tags on all records as they are created to add
them to the PVI of the controller.
Add call to add sub controller PVI to parent PVI